+++ src/app/page.tsx
@@ -6,8 +6,34 @@ import { fetchAuction } from '@/services/auction';
 import { DAO_ADDRESSES } from '@/utils/constants';
 import { Grid, GridItem, VStack } from '@chakra-ui/react';
 import { headers } from 'next/headers';
+import type { Metadata } from 'next';
 
 export const revalidate = 60;
+
+const appUrl = process.env.NEXT_PUBLIC_URL;
+
+const frame = {
+  version: 'next',
+  imageUrl: `${appUrl}/frames/auction/opengraph-image`,
+  button: {
+    title: 'Gnars Auction',
+    action: {
+      type: 'launch_frame',
+      name: 'Gnars Auction',
+      url: `${appUrl}/frames/auction/`,
+      splashImageUrl: `${appUrl}/images/loading.gif`,
+      splashBackgroundColor: '#f7f7f7',
+    },
+  },
+};
+
+export const metadata: Metadata = {
+  title: 'Gnars Dao',
+  description: 'Gnarly Ecosystem',
+  other: {
+    'fc:frame': JSON.stringify(frame),
+  },
+};
 
 async function App() {
   headers();
